+++ CHANGELOG.md
@@ -5,10 +5,13 @@ This change log adheres to standards from [Keep a CHANGELOG](https://keepachange
 
 ## Unreleased
 
+## [7.32.1] - 2023.01.16
+
 ### Fixed
 * prevent circular dependency in index and "all" config ([#3519][] @ljharb)
 * [`destructuring-assignment`]: do not force destructuring of optionally chained properties ([#3520][] @ljharb)
 
+[7.32.1]: https://github.com/jsx-eslint/eslint-plugin-react/compare/v7.32.0...v7.32.1
 [#3520]: https://github.com/jsx-eslint/eslint-plugin-react/issues/3520
 [#3519]: https://github.com/jsx-eslint/eslint-plugin-react/issues/3519